% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.exists.R
\name{all_exists}
\alias{all_exists}
\title{All Files exist}
\usage{
all_exists(...)
}
\arguments{
\item{...}{files to be passed to \code{\link{file.exists}}}
}
\value{
Logical length of \code{...}
}
\description{
Wrapper for all files existing
}
