% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_mask.R
\name{apply_mask}
\alias{apply_mask}
\title{Apply Brain mask}
\usage{
apply_mask(x, mask, outdir = tempdir(), verbose = TRUE,
  suffix = "_brain")
}
\arguments{
\item{x}{List of images or character vector.  Must be named with
the imaging modalities.}

\item{mask}{Brain Mask image/filename}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{suffix}{Name to append to the image filename}
}
\value{
List of masked brains
}
\description{
Apply Brain mask
}
