% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_spatial_normalize.R
\name{apply_spatial_normalize}
\alias{apply_spatial_normalize}
\title{Apply Spatial Normalization}
\usage{
apply_spatial_normalize(x, template = c("none", "Eve", "MNI"),
  template_fname, fwdtransforms = NULL, suffix = "",
  interpolator = "lanczosWindowedSinc", outdir = tempdir(),
  verbose = TRUE, copy_origin = TRUE)
}
\arguments{
\item{x}{List of filenames}

\item{template}{Resampling to \code{1x1x1} (template = none) or
Template to register to - either MNI or Eve template}

\item{template_fname}{Filename of Template}

\item{fwdtransforms}{Transforms to template if not resampling}

\item{suffix}{Name to append to the image filename}

\item{interpolator}{interpolation passed to \code{\link{ants_apply_transforms}}}

\item{outdir}{Output directory}

\item{verbose}{Print diagnostic messages}

\item{copy_origin}{Copy image origin from \code{target},
using \code{\link{antsCopyOrigin}}}
}
\value{
List of filenames of normalized data
}
\description{
Apply Spatial Normalization
}
