% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_pauc.R
\name{auc_pauc}
\alias{auc_pauc}
\title{Get the AUC and pAUC for a prediction}
\usage{
auc_pauc(prediction.obj, fpr.stop = 0.01)
}
\arguments{
\item{prediction.obj}{An object of class \code{\link{prediction}}.}

\item{fpr.stop}{Limit of false positive rate.}
}
\value{
Depending on the number of predictions, either a vector of length 2 or
a matrix of 2 columns, with the first column being the AUC, the second - pAUC
}
\description{
Get the area under the curve (AUC) and the partial AUC (pAUC)
for a prediction object
}
