% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_noneck_reduce.R
\name{bc_noneck_reduce}
\alias{bc_noneck_reduce}
\title{Bias Correct, Remove Neck, Reduce Image}
\usage{
bc_noneck_reduce(x, remove_negative = TRUE, gold_standard = NULL,
  gs_space = NULL, reduce = TRUE, cleanup = 1, probs = c(0, 0.999),
  non_zero = TRUE, outdir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{List of images or character vector.  Must be named with
the imaging modalities.}

\item{remove_negative}{Force values less than zero to zero}

\item{gold_standard}{Gold Standard image/filename, if applicable}

\item{gs_space}{space the Gold Standard is located}

\item{reduce}{If \code{TRUE}, the image size will be reduced using
\code{\link{reduce_img}}.}

\item{cleanup}{Argument to \code{\link{oMask}}.  If >0, morphological
operations will be applied to clean up the mask by
eroding away small or weakly-connected areas, and closing holes.}

\item{probs}{passed to \code{\link{winsor}} for Winsorization}

\item{non_zero}{Should zeroes be excluded from the calculation
of quantiles? Passed to \code{\link{winsor}}.}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}
}
\value{
List of output filenames
}
\description{
Bias Correct, Remove Neck, Reduce Image
}
