% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cutoff}
\alias{cutoff}
\title{Cutoff for random forest model}
\format{An object of class \code{numeric} of length 1.}
\usage{
cutoff
}
\description{
Cutoff for random forest model
}
\keyword{datasets}
