% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Calculate Dice Coefficient}
\usage{
dice(prediction.obj)
}
\arguments{
\item{prediction.obj}{object of class \code{\link{prediction-class}} from
\code{\link{prediction}}}
}
\value{
Object of class \code{\link{performance}}
}
\description{
Calculate Dice Coefficient/Similarity Index from a prediction object
}
