% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsl_smooth_images.R
\name{fsl_smooth_images}
\alias{fsl_smooth_images}
\title{Smoothing sMRI data with FSL}
\usage{
fsl_smooth_images(x, sigma = c(3, 5, 10, 20), mask = NULL,
  outdir = tempdir(), suffix = "_smooth", verbose = TRUE, ...)
}
\arguments{
\item{x}{List of images}

\item{sigma}{vector of sigmas for smoothers in FSL}

\item{mask}{brain mask for smoother}

\item{outdir}{output directory}

\item{suffix}{Name to append to the image filename.  The sigma will
be appended to this suffix.}

\item{verbose}{print diagnostic messages}

\item{...}{Additional arguments to \code{\link{fsl_smooth}}}
}
\value{
List of the images, brain mask, suffix, and output directory
}
\description{
Smoothing sMRI data with FSL
}
