% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_dice.R
\name{get_max_dice}
\alias{get_max_dice}
\title{Get the Maximum Dice Coefficient and Cutoff}
\usage{
get_max_dice(prediction.obj)
}
\arguments{
\item{prediction.obj}{An object of class \code{\link{prediction}}.}
}
\value{
Depending on the number of predictions, either a vector of length 2 or
a matrix of 2 columns, with the first column being the dice, the second - cutoff
}
\description{
Get the Dice coefficient and the cutoff for that max Dice
from a prediction object
}
