% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity_normalize.R
\name{intensity_normalize}
\alias{intensity_normalize}
\title{SMRI Normalize}
\usage{
intensity_normalize(x, mask = NULL, normalization = c("z", "trimmed_z",
  "quantile", "minmax", "whitestripe_T1", "whitestripe_T2",
  "whitestripe_hybrid"), outdir = tempdir(), verbose = TRUE,
  remask = TRUE, suffix = "_z", ...)
}
\arguments{
\item{x}{List of images}

\item{mask}{mask to normalize over}

\item{normalization}{Normalization Method}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{remask}{should mask_img be applied to the output images
after normalization}

\item{suffix}{what the append to end of the output filename}

\item{...}{Additional arguments to \code{normaliztaion} function}
}
\value{
List of images
}
\description{
SMRI Normalize
}
