% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df.R
\name{make_df}
\alias{make_df}
\title{Make Data.frame of images}
\usage{
make_df(fnames, mask, y_fname = NULL, verbose = TRUE)
}
\arguments{
\item{fnames}{Filenames to put into data.frame}

\item{mask}{Mask to subset the images}

\item{y_fname}{Name of gold standard}

\item{verbose}{print diagnostic messages}
}
\value{
List of mask and data.frame
}
\description{
Take the filenames and make an analytic data.frame
}
