% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_processed_filenames.R
\name{make_processed_filenames}
\alias{make_processed_filenames}
\title{Make Processed Filenames}
\usage{
make_processed_filenames(types = c("FLAIR", "T1_Pre", "T1_Post", "T2", "PD"),
  gold_standard = TRUE, outdir = ".")
}
\arguments{
\item{types}{Images used}

\item{gold_standard}{Should the gold standard be added}

\item{outdir}{output directory of processing}
}
\value{
data.frame of values
}
\description{
Get filename output from processing
}
