% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malf_tissue_seg.R
\name{malf_tissue_seg}
\alias{malf_tissue_seg}
\title{MALF tissue Segmentation}
\usage{
malf_tissue_seg(t1, outfile = NULL, num_templates = 15,
  interpolator = "NearestNeighbor", typeofTransform = "SyN",
  func = "mode", keep_regs = TRUE, verbose = TRUE)
}
\arguments{
\item{t1}{T1-weighted image/filename}

\item{outfile}{Output filename}

\item{num_templates}{Number of templates to use for MALF}

\item{interpolator}{Interpolation used}

\item{typeofTransform}{Transformation to align the templates to the T1}

\item{func}{Function used to vote}

\item{keep_regs}{Keep registrations (for future use),
passed to \code{\link{malf}}}

\item{verbose}{Print diagnostic messages}
}
\value{
Object of class nifti
}
\description{
Uses MALF for tissue-class segmentation
}
