% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_normalize.R
\name{minmax}
\alias{minmax}
\title{Normalize using Min/Max}
\usage{
minmax(img, mask = NULL, remask = TRUE)
}
\arguments{
\item{img}{Image to normalize}

\item{mask}{Mask same dimensions as image}

\item{remask}{Should the image be remasked after normalizing?}
}
\value{
Object of same class as img
}
\description{
Normalize using Min/Max
}
