% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msseg_pipeline.R
\name{msseg_pipeline}
\alias{msseg_pipeline}
\title{MSSeg Pipeline}
\usage{
msseg_pipeline(t1_pre = "3DT1.nii.gz", t1_post = "3DT1GADO.nii.gz",
  flair = "3DFLAIR.nii.gz", t2 = "T2.nii.gz", pd = "DP.nii.gz",
  gold_standard = NULL, outdir = ".", outfile = NULL,
  num_templates = 15, verbose = TRUE, runcc = TRUE, min_vol = 0.01,
  remove_files = FALSE, ...)
}
\arguments{
\item{t1_pre}{T1-weighted pre-gad image/filename}

\item{t1_post}{T1-weighted post-gad image/filename}

\item{flair}{FLAIR image/filename}

\item{t2}{T2-weighted image/filename}

\item{pd}{Proton Density pre-gad image/filename}

\item{gold_standard}{Gold Standar Lesion}

\item{outdir}{Output directory}

\item{outfile}{Output file for segmentation}

\item{num_templates}{Number of templates used in MASS}

\item{verbose}{print diagnostic messages}

\item{runcc}{Run connected components}

\item{min_vol}{Minimum volume of a lesion}

\item{remove_files}{Remove predictor files from hard disk}

\item{...}{arguments passed to \code{\link{process_images}}}
}
\value{
Final Segmentation
}
\description{
Full Pipeline for MSSeg Challenge
}
