% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_fast.R
\name{multi_fast}
\alias{multi_fast}
\title{Multiple FAST for PD/T1/T2}
\usage{
multi_fast(x, outdir = tempdir(), bias_correct = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{x}{List of processed images with T1/T2/PD}

\item{outdir}{Output directory}

\item{bias_correct}{should bias correction be done?  passed to
\code{\link{fast}}}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments to \code{\link{fast}}}
}
\value{
List of filenames
}
\description{
Multiple FAST for PD/T1/T2
}
