% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n4.R
\name{n4_raw}
\alias{n4_raw}
\title{N4 Bias Field Correction}
\usage{
n4_raw(x, verbose = TRUE, mask = NULL, outdir = tempdir(),
  suffix = "_N4", convergence = list(iters = rep(200L, 4), tol = 1e-07))
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{mask}{Mask for Bias-field correction}

\item{outdir}{Output directory}

\item{suffix}{Name to append to the image filename}

\item{convergence}{convergence parameters for
\code{\link{n4BiasFieldCorrection}}}
}
\value{
List of filenames
}
\description{
N4 Bias Field Correction
}
