% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noneck.R
\name{noneck}
\alias{noneck}
\title{Remove Neck (Doubly)}
\usage{
noneck(x, verbose = TRUE, outdir = tempdir())
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{outdir}{Output directory}
}
\value{
List of filenames
}
\description{
Remove Neck (Doubly)
}
