% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pred.R
\name{norm_predictors}
\alias{norm_predictors}
\title{SMRI Intensity Normalize and Create Predictors}
\usage{
norm_predictors(normalized, normalization = c("trimmed_z", "z",
  "quantile", "minmax", "whitestripe_T1", "whitestripe_T2",
  "whitestripe_hybrid"), remask = TRUE, suffix = normalized$suffix,
  outdir = normalized$outdir, verbose = TRUE, radius = rep(1, 3),
  sigma = c(3, 5, 10, 20), ...)
}
\arguments{
\item{normalized}{The \code{normalized} element from
\code{\link{seg_normalize}} output}

\item{normalization}{Normalization Method}

\item{remask}{should mask_img be applied to the output images
after normalization}

\item{suffix}{what the append to end of the output filename}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{radius}{radius of neighborhood.  See \code{\link{smri_corr_img}}
and \code{\link{smri_moments}}}

\item{sigma}{vector of sigmas for smoothers in FSL see
\code{\link{fsl_smooth_images}}}

\item{...}{Additional arguments to \code{normaliztaion} function}
}
\value{
List of images
}
\description{
SMRI Intensity Normalize and Create Predictors
}
