% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_converter.R
\name{png_converter}
\alias{png_converter}
\title{Convert set of PNGs to PDF}
\usage{
png_converter(pngs, pdfname, extra.opts = "-density 300", interval = 0,
  autobrowse = FALSE)
}
\arguments{
\item{pngs}{List of names (or with wildcards) of png files}

\item{pdfname}{Output PDF filename}

\item{extra.opts}{Options passed to \code{\link{im.convert}}}

\item{interval}{interval of delay, passed to \code{\link{ani.options}}}

\item{autobrowse}{passed to \code{\link{ani.options}}}
}
\value{
Name of PDF output file
}
\description{
Converts PNGs to PDF
}
