% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_images}
\alias{process_images}
\title{Process Images for MSSeg}
\usage{
process_images(t1_pre, t1_post, flair, t2, pd, gold_standard = NULL,
  outdir = ".", num_templates = 15, verbose = TRUE, force = TRUE,
  cleanup = TRUE)
}
\arguments{
\item{t1_pre}{T1-weighted pre-gad image/filename}

\item{t1_post}{T1-weighted post-gad image/filename}

\item{flair}{FLAIR image/filename}

\item{t2}{T2-weighted image/filename}

\item{pd}{Proton Density pre-gad image/filename}

\item{gold_standard}{Gold Standar Lesion}

\item{outdir}{Output directory}

\item{num_templates}{Number of templates used in MASS}

\item{verbose}{print diagnostic messages}

\item{force}{Force tissue segmentations}

\item{cleanup}{delete temporary files}
}
\value{
List of output filenames
}
\description{
Processing script for MSSeg Challenge
}
