% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_mask.R
\name{quantile_mask}
\alias{quantile_mask}
\title{Threshold Quantile Image}
\usage{
quantile_mask(img, mask, probs = 0.85)
}
\arguments{
\item{img}{Character vector, or object of class \code{nifti}}

\item{mask}{Mask to determine cumulative distribution function (cdf) from}

\item{probs}{Probability to threshold at}
}
\value{
Object of class \code{\link{nifti}}
}
\description{
Threshold Quantile Image
}
