% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reapply_malf_tissue_seg.R
\name{reapply_malf_tissue_seg}
\alias{reapply_malf_tissue_seg}
\title{MALF tissue Segmentation without Re-registration}
\usage{
reapply_malf_tissue_seg(t1, regs, outfile = NULL, num_templates = 15,
  interpolator = "NearestNeighbor", typeofTransform = "SyN",
  func = "mode", verbose = TRUE, ...)
}
\arguments{
\item{t1}{T1-weighted image/filename}

\item{regs}{List of registrations from malf or malf_registration,
each element must have fwdtransforms and interpolator.
Length must match \code{num_templates}.}

\item{outfile}{Output filename}

\item{num_templates}{Number of templates to use for MALF}

\item{interpolator}{Interpolation used}

\item{typeofTransform}{Transformation to align the templates to the T1}

\item{func}{Function used to vote}

\item{verbose}{Print diagnostic messages}

\item{...}{additional options to pass to \code{reapply_malf}}
}
\value{
Object of class nifti
}
\description{
Uses MALF for tissue-class segmentation, but with already-done
registrations
}
