% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_img.R
\name{reduce_img}
\alias{reduce_img}
\alias{mask_reduce}
\title{Reduce Image, Dropping Image Dimensions}
\usage{
reduce_img(x, verbose = TRUE, cleanup = 1, outdir = tempdir(),
  suffix = "_reduced")

mask_reduce(x, cleanup = 1)
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{cleanup}{Argument to \code{\link{oMask}}.  If >0, morphological
operations will be applied to clean up the mask by
eroding away small or weakly-connected areas, and closing holes.}

\item{outdir}{Output directory}

\item{suffix}{Name to append to the image filename}
}
\value{
List of filenames
}
\description{
Reduce Image, Dropping Image Dimensions
}
