% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_model_1}
\alias{rf_model_1}
\title{Random forest model 1}
\format{An object of class \code{randomForest} of length 13.}
\usage{
rf_model_1
}
\description{
Random forest model 1
}
\keyword{datasets}
