% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_model_2}
\alias{rf_model_2}
\title{Random forest model 2}
\format{An object of class \code{randomForest} of length 13.}
\usage{
rf_model_2
}
\description{
Random forest model 2
}
\keyword{datasets}
