% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rf_model_9}
\alias{rf_model_9}
\title{Random forest model}
\format{An object of class \code{randomForest} of length 13.}
\usage{
rf_model_9
}
\description{
Random forest model
}
\keyword{datasets}
