% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpi_raw.R
\name{rpi_raw}
\alias{rpi_raw}
\title{RPI Orientation of Images}
\usage{
rpi_raw(x, gold_standard = NULL, verbose = TRUE)
}
\arguments{
\item{x}{List of images or character vector.  Must be named with
the imaging modalities. T1 and FLAIR must be included}

\item{gold_standard}{Gold Standard image/filename, if applicable}

\item{verbose}{print diagnostic messages}
}
\value{
List of reoriented output filenames
}
\description{
RPI Orientation of Images
}
