% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_normalize.R
\name{seg_normalize}
\alias{seg_normalize}
\title{Spatially Normalize SMRI data and Tissue Segmentation}
\usage{
seg_normalize(prenormalize, template = c("none", "Eve", "MNI"),
  norm_outdir = prenormalize$outdir, verbose = TRUE,
  typeofTransform = "SyN", segment_typeofTransform = typeofTransform,
  interpolator = "lanczosWindowedSinc",
  dis_interpolator = "genericLabel",
  malf_interpolator = "genericLabel", copy_origin = TRUE,
  force_registration = FALSE, ...)
}
\arguments{
\item{prenormalize}{list from \code{\link{smri_prenormalize}}}

\item{template}{Resampling to \code{1x1x1} (template = none) or
Template to register to - either MNI or Eve template}

\item{norm_outdir}{Output directory of spatially normalized data.  The
output directory of the segmentation images will still be specified
in \code{prenormalize$outdir}}

\item{verbose}{print diagnostic messages}

\item{typeofTransform}{Transformation to use for registration,
passed to \code{\link{registration}}}

\item{segment_typeofTransform}{Transformation to use for
\code{malf} segmentation}

\item{interpolator}{Interpolation done, passed to
\code{\link{registration}} for continuous data}

\item{dis_interpolator}{Interpolation done, passed to
\code{\link{registration}} for discrete data}

\item{copy_origin}{Copy image origin from image
being registered, using \code{\link{antsCopyOrigin}}}

\item{force_registration}{Should the MALF registrations be re-run if
they already exist?}

\item{...}{arguments passed to \code{\link{t1_segment}}}
}
\value{
List of images, suffix, brain mask, gold standard,
and registration if applicable
}
\description{
Spatially Normalize SMRI data and Tissue Segmentation, note T1 is
required
}
