% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_origin_zero.R
\name{set_origin_zero}
\alias{set_origin_zero}
\title{Set Origin to Zero for the Images}
\usage{
set_origin_zero(fnames)
}
\arguments{
\item{fnames}{character vector of filenames}
}
\value{
Logical indicating if the setting was successful
}
\description{
Set Origin to Zero for the Images
}
