% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_corr_img.R
\name{smri_corr_img}
\alias{smri_corr_img}
\title{SMRI Local Correlation Images}
\usage{
smri_corr_img(x, mask = NULL, outdir = tempdir(), verbose = TRUE,
  suffix = "_corr", radius = rep(1, 3), method = c("pearson",
  "spearman"))
}
\arguments{
\item{x}{List of images}

\item{mask}{mask to normalize over}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{suffix}{what the append to end of the output filename}

\item{radius}{radius of neighborhood.  See \code{\link{corr_img}}}

\item{method}{Method for correlation. See \code{\link{corr_img}}}
}
\value{
List of filenames
}
\description{
Calculates the local correlation between sets of images
}
