% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_flipped.R
\name{smri_flipped}
\alias{smri_flipped}
\title{SMRI Flipped Image (x-y flipped)}
\usage{
smri_flipped(x, mask = NULL, outdir = tempdir(), verbose = TRUE,
  suffix = "_flip", typeofTransform = "SyN", interpolator = "Linear")
}
\arguments{
\item{x}{List of images}

\item{mask}{mask to normalize over}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{suffix}{what the append to end of the output filename}

\item{typeofTransform}{type of transformed used, passed to
\code{\link{registration}}}

\item{interpolator}{interpolation done in \code{\link{registration}}}
}
\value{
List of filenames
}
\description{
Flips an image and registers the image to the original.
Assumes RL orientation
}
