% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_moments.R
\name{smri_moments}
\alias{smri_moments}
\title{Structural MRI Moment images}
\usage{
smri_moments(x, mask = NULL, radius = rep(1, 3), suffix = "_",
  outdir = tempdir(), verbose = TRUE, moments = c("mn", "sd", "skew",
  "kurt", "grad", "z"), ...)
}
\arguments{
\item{x}{List of images}

\item{mask}{brain mask for moment}

\item{radius}{array of values for neighborhood radius (in voxels),
passed to \code{\link{neighborhood}}}

\item{suffix}{Name to append to the image filename.  The moments will
be appended to this suffix.  Should likely end with an underscore (_) or
other separator}

\item{outdir}{output directory}

\item{verbose}{print diagnostic messages}

\item{moments}{Moments to create, mean, standard deviation,
skew, kurtosis, gradient, z-score (mean/sd)}

\item{...}{Additional arguments to \code{\link{create_moment}}}
}
\value{
A list of filenames, separated by imaging modality, containing
filenames for all the moments specified
}
\description{
Structural MRI Moment images
}
