% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_normalize.R
\name{smri_normalize}
\alias{smri_normalize}
\title{SMRI Normalize}
\usage{
smri_normalize(x, mask = NULL, normalization = c("z", "trimmed_z",
  "quantile", "minmax"), outdir = tempdir(), verbose = TRUE, ...)
}
\arguments{
\item{x}{List of images}

\item{mask}{mask to normalize over}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{...}{Additional arguments to MALF}
}
\value{
List of images
}
\description{
SMRI Normalize
}
