% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_normalize.R
\name{spatial_normalize}
\alias{spatial_normalize}
\title{Spatially Normalize SMRI data}
\usage{
spatial_normalize(prenormalize, template = c("none", "Eve", "MNI"),
  verbose = TRUE, typeofTransform = "SyN",
  interpolator = "lanczosWindowedSinc", dis_interpolator = "genericLabel")
}
\arguments{
\item{prenormalize}{list from \code{\link{smri_prenormalize}}}

\item{template}{Resampling to \code{1x1x1} (template = none) or
Template to register to - either MNI or Eve template}

\item{verbose}{print diagnostic messages}

\item{typeofTransform}{Transformation to use for registration,
passed to \code{\link{registration}}}

\item{interpolator}{Interpolation done, passed to
\code{\link{registration}}}

\item{dis_interpolator}{Interpolation done, passed to
\code{\link{registration}} for discrete data}
}
\value{
List of images, suffix, brain mask, gold standard,
and registration if applicable
}
\description{
Spatially Normalize SMRI data
}
