% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_model.R
\name{strip_model}
\alias{strip_model}
\title{Strip Model to Reduce Memory Footprint}
\usage{
strip_model(model)
}
\arguments{
\item{model}{Object of class \code{glm} or \code{lm}}
}
\value{
Object of class \code{glm} or \code{lm}, but with some elements removed
}
\description{
Strips elements from a \code{glm} or \code{lm} object to reduce size.
Allows for printing/showing, but \code{summary} must be run before stripping
}
\examples{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)
object.size(lm.D9)
object.size(strip_model(lm.D9))
}
