% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_segment.R
\name{t1_segment}
\alias{t1_segment}
\title{MALF Tissue and Structure Segmentation}
\usage{
t1_segment(t1, outdir = tempdir(), num_templates = 35,
  interpolator = "genericLabel", typeofTransform = "SyNAggro",
  verbose = TRUE, malf_prob_function = "staple_prob",
  malf_label_function = "mode", rerun_malf = FALSE, ...)
}
\arguments{
\item{t1}{T1-weighted image/filename}

\item{outdir}{Output directory}

\item{num_templates}{Number of templates to use for MALF}

\item{interpolator}{interpolation done in \code{\link{registration}}}

\item{typeofTransform}{Transformation to align the templates to the T1}

\item{verbose}{Print diagnostic messages}

\item{malf_prob_function}{Function for MALF combination for probability.}

\item{malf_label_function}{Function for MALF combination for labeling.}

\item{rerun_malf}{Should the MALF registrations be re-run if
they already exist?}

\item{...}{additional options to pass to \code{\link{malf}}}
}
\value{
Object of class nifti
}
\description{
MALF Tissue and Structure Segmentation
}
