% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_mean_sd.R
\name{trim_mean_sd}
\alias{trim_mean_sd}
\title{Trimmed Mean and Standard Deviation}
\usage{
trim_mean_sd(x, trim = 0.2)
}
\arguments{
\item{x}{numeric vector}

\item{trim}{value between 0 and 0.5, percent trim from each side of the distribution.
Calculates quantiles for \code{trim} and \code{1-trim} and sets values outside to NA}
}
\value{
Vector of length 2, with mean/sd named elements
}
\description{
Calculates both trimmed mean and standard deviation from a vector
}
\examples{
x = rnorm(100)
trim_mean_sd(x)
}
