% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean.R
\name{trimmed_ratio}
\alias{trimmed_ratio}
\title{Trimmed Ratio Image}
\usage{
trimmed_ratio(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{trimmed_z}}}
}
\value{
Object of class \code{nifti}
}
\description{
Creates a normalized image using the image divided by a
trimmed standard deviation, but not centered
}
\examples{
nim = oro.nifti::nifti(img = array(rnorm(1000), dim = c(10, 10, 10)))
mask = nim > 0
trimmed_z(nim, mask)
}
