% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmean.R
\name{trimmed_z}
\alias{trimmed_z}
\title{Trimmed Z-score Image}
\usage{
trimmed_z(img, mask = NULL, remask = FALSE, trim = 0.2, center = TRUE,
  scale = TRUE)
}
\arguments{
\item{img}{Object of class \code{nifti}}

\item{mask}{Binary object of class \code{nifti}.  If missing, whole image used.}

\item{remask}{Should the result be masked after z-scoring (all outside set to 0)}

\item{trim}{The trim passed to \code{\link{trim_mean_sd}}}

\item{center}{Logical indicating whether the value should be centered by
trimmed mean}

\item{scale}{Logical indicating whether the value should be scaled by
trimmed sd}
}
\value{
Object of class \code{nifti}
}
\description{
Creates a normalized image using z-scores
}
\examples{
nim = oro.nifti::nifti(img = array(rnorm(1000), dim = c(10, 10, 10)))
mask = nim > 0
trimmed_z(nim, mask)
}
