% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstrip_image.R
\name{unstrip_image}
\alias{unstrip_image}
\title{Replace Dropped Dimensions from Stripping}
\usage{
unstrip_image(strip_image, outdir = ".")
}
\arguments{
\item{strip_image}{image with dropped dimensions}

\item{outdir}{Location of \code{FLAIR_MALF_Brain_Mask.nii.gz} and
\code{FLAIR_noneck.nii.gz}}
}
\value{
Nifti object of same dimensions as FLAIR
}
\description{
Takes in a _strip segmentation and makes it same size as FLAIR
}
