% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_images.R
\name{z_images}
\alias{z_images}
\title{Standardized to Template Images SyN}
\usage{
z_images(orig_t1_pre, t1_pre, t1_post, flair, t2, pd, outdir = ".",
  add_stub = "_strip_norm")
}
\arguments{
\item{orig_t1_pre}{Skull-stripped T1 image to register to Eve}

\item{t1_pre}{Normalized T1 Pre}

\item{t1_post}{Normalized T1 Post}

\item{flair}{Normalized FLAIR}

\item{t2}{Normalized T2}

\item{pd}{Normalized PD}

\item{outdir}{output directory}

\item{add_stub}{Stub to add on the filenames}
}
\description{
Normalizes to the template
}
