% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_noneck_reduce.R
\name{bc_noneck_reduce}
\alias{bc_noneck_reduce}
\title{Bias Correct, Remove Neck, Reduce Image}
\usage{
bc_noneck_reduce(x, gold_standard = NULL, gs_space = NULL,
  outdir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{List of images or character vector.  Must be named with
the imaging modalities. T1 and FLAIR must be included}

\item{gold_standard}{Gold Standard Lesion}

\item{gs_space}{space the Gold Standard is located}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}
}
\value{
List of output filenames
}
\description{
Bias Correct, Remove Neck, Reduce Image
}
