% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n4.R
\name{n4_raw}
\alias{n4_raw}
\title{N4 Bias Field Correction}
\usage{
n4_raw(x, verbose = TRUE, outdir = tempdir())
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{outdir}{Output directory}
}
\value{
List of filenames
}
\description{
N4 Bias Field Correction
}
