% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_img.R
\name{reduce_img}
\alias{reduce_img}
\alias{mask_reduce}
\title{Reduce Image, Dropping Image Dimensions}
\usage{
reduce_img(x, verbose = TRUE, outdir = tempdir(), suffix = "_reduced")

mask_reduce(x)
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{outdir}{Output directory}

\item{suffix}{Name to append to the image filename}
}
\value{
List of filenames
}
\description{
Reduce Image, Dropping Image Dimensions
}
