% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_to_t1.R
\name{reg_to_t1}
\alias{reg_to_t1}
\title{Register Images to Another image (usually T1)}
\usage{
reg_to_t1(x, gs_space = NULL, interpolator = "Linear",
  gs_interpolator = "NearestNeighbor", outdir = tempdir(), verbose = TRUE,
  reg_space = "T1", suffix = "_regtoT1", remove_negative = TRUE)
}
\arguments{
\item{x}{List of images or character vector.  Must be named with
the imaging modalities. The \code{reg_space} file must be included}

\item{gs_space}{space the Gold Standard is located (if applicable)}

\item{interpolator}{interpolation done in \code{\link{registration}}}

\item{gs_interpolator}{interpolation done in \code{\link{registration}}
for gold standard}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{reg_space}{space to register images to}

\item{suffix}{Name to append to the image filename}

\item{remove_negative}{After registration, any values < 0 are set to 0}
}
\value{
List of Images
}
\description{
Register Images to Another image (usually T1)
}
