% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_prenormalize.R
\name{smri_prenormalize}
\alias{smri_prenormalize}
\title{SMRI Pipeline for Preprocessing Prior to Normalization}
\usage{
smri_prenormalize(x, gold_standard = NULL, gs_space = NULL,
  probs = c(0, 0.995), non_zero = TRUE,
  interpolator = "lanczosWindowedSinc", brain_mask = NULL,
  gs_interpolator = "genericLabel", num_templates = 15,
  malf_transform = "SyNAggro", outdir = tempdir(), reg_space = "T1",
  brain_extraction_method = c("abp", "malf"),
  brain_malf_function = "staple_prob", n_skull_iter = 3,
  brain_threshold = 0.5, verbose = TRUE, remove_negative = TRUE,
  zero_origin = TRUE, reduce = TRUE, cleanup = 1, ...)
}
\arguments{
\item{x}{List of images}

\item{gold_standard}{Gold Standard image/filename, if applicable}

\item{gs_space}{space the Gold Standard is located}

\item{probs}{passed to \code{\link{winsor}} for Winsorization}

\item{non_zero}{Should zeroes be excluded from the calculation
of quantiles? Passed to \code{\link{winsor}}.}

\item{interpolator}{interpolation passed to \code{\link{reg_to_t1}}}

\item{brain_mask}{Brain Mask in \code{reg_space}}

\item{gs_interpolator}{Gold Standard interpolation
passed to \code{\link{reg_to_t1}}}

\item{num_templates}{Number of templates to use for MALF}

\item{malf_transform}{type of registration transformation for MALF}

\item{outdir}{Output directory}

\item{reg_space}{space to register images to}

\item{brain_extraction_method}{Brain extraction method, either
\code{\link{malf}} or \code{abp} for \code{\link{n4_skull_strip}}}

\item{brain_malf_function}{Function to be passed to \code{\link{malf}},
and subsequently \code{\link{stat_img}} for brain extraction}

\item{n_skull_iter}{number of iterations to use
\code{\link{n4_skull_strip}} for skull stripping if
\code{brain_extraction_method = "abp"}.}

\item{brain_threshold}{Percentage threshold for the brain MALF image
to be considered part of the mask}

\item{verbose}{print diagnostic messages}

\item{remove_negative}{After reorientation, N4, and registration,
any values < 0 are set to 0}

\item{zero_origin}{Should the origin be set to 0 for the
image to be registered to?}

\item{reduce}{If \code{TRUE}, the image size will be reduced using
\code{\link{reduce_img}}.}

\item{cleanup}{Argument to \code{\link{reduce_img}} for reducing the images.}

\item{...}{Additional arguments to MALF}
}
\value{
List of the images, brain mask, suffix, and output directory
}
\description{
SMRI Pipeline for Preprocessing Prior to Normalization
}
