% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smri_prenormalize.R
\name{smri_prenormalize}
\alias{smri_prenormalize}
\title{SMRI Pipeline for Preprocessing Prior to Normalization}
\usage{
smri_prenormalize(x, gold_standard = NULL, gs_space = NULL, probs = c(0,
  0.999), interpolator = "lanczosWindowedSinc", brain_mask = NULL,
  gs_interpolator = "NearestNeighbor", num_templates = 15,
  malf_transform = "SyNAggro", outdir = tempdir(), verbose = TRUE, ...)
}
\arguments{
\item{x}{List of images}

\item{gold_standard}{Gold Standard image/filename, if applicable}

\item{gs_space}{space the Gold Standard is located}

\item{probs}{passed to \code{\link{winsor}} for Winsorization}

\item{interpolator}{interpolation passed to \code{\link{reg_to_t1}}}

\item{brain_mask}{Brain Mask in T1 Space}

\item{gs_interpolator}{Gold Standard interpolation
passed to \code{\link{reg_to_t1}}}

\item{num_templates}{Number of templates to use for MALF}

\item{malf_transform}{type of registration transformation for MALF}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{...}{Additional arguments to MALF}
}
\value{
List of the images, brain mask, suffix, and output directory
}
\description{
SMRI Pipeline for Preprocessing Prior to Normalization
}
