% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_segment.R
\name{t1_segment}
\alias{t1_segment}
\title{MALF Tissue and Structure Segmentation}
\usage{
t1_segment(t1, outdir = tempdir(), num_templates = 35,
  interpolator = "genericLabel", typeofTransform = "SyNAggro",
  verbose = TRUE, ...)
}
\arguments{
\item{t1}{T1-weighted image/filename}

\item{outdir}{Output directory}

\item{num_templates}{Number of templates to use for MALF}

\item{interpolator}{interpolation done in \code{\link{registration}}}

\item{typeofTransform}{Transformation to align the templates to the T1}

\item{verbose}{Print diagnostic messages}

\item{...}{additional options to pass to \code{reapply_malf}}
}
\value{
Object of class nifti
}
