% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_z_score.R
\name{template_z_score}
\alias{template_z_score}
\title{SMRI Z-score compared to a Template}
\usage{
template_z_score(x, mask = NULL, template = "Eve", mean_imgs = NULL,
  sd_imgs = NULL, outdir = tempdir(), verbose = TRUE,
  remask = TRUE, lower_value = -20, upper_value = 20,
  add_const = TRUE, interpolator = "lanczosWindowedSinc",
  suffix = "_ztemp")
}
\arguments{
\item{x}{List of images}

\item{mask}{mask to place all non-mask values to zero}

\item{template}{Template of the space of the mean/SD images,
only currently accepts Eve.}

\item{mean_imgs}{A vector/list of images for measures of
centrality - usually the mean.
See \code{\link{download_template_img_data}}}

\item{sd_imgs}{A vector/list of images for measures of
spread - usually the standard deviation.
See \code{\link{download_template_img_data}}#'}

\item{outdir}{Output directory}

\item{verbose}{print diagnostic messages}

\item{remask}{should \code{\link{mask_img}}
be applied to the output images
after normalization}

\item{lower_value}{Lower value to winsorize the z-score to.
With small standard deviations, the values can get far outside of
a standard range.}

\item{upper_value}{upper value to winsorize the z-score to.
With small standard deviations, the values can get far outside of
a standard range.}

\item{add_const}{Should a constant be added
to the T1 image; without this, the registration may
fail}

\item{interpolator}{Interpolation done, passed to
\code{\link{registration}}}

\item{suffix}{what the append to end of the output filename}
}
\value{
List of images
}
\description{
SMRI Z-score compared to a Template
}
