% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{Winsorize Image}
\usage{
winsor(x, verbose = TRUE, outdir = tempdir(), probs = c(0, 0.999),
  suffix = "_winsor")
}
\arguments{
\item{x}{List of processed filenames}

\item{verbose}{print diagnostic messages}

\item{outdir}{Output directory}

\item{probs}{passed to \code{\link{robust_window}} for Winsorization}

\item{suffix}{Name to append to the image filename}
}
\value{
List of filenames
}
\description{
Winsorize Image
}
